requirements.txt
python-dotenv
requests
anthropic
psycopg2-binary
bcrypt
cryptography
